requirements.txt
hydra-core==1.3.1
numpy==1.21.5
omegaconf==2.2.3
pandas==1.2.3
pymoo==0.6.0
pytorch_warmup==0.1.1
PyYAML==6.0
scipy==1.3.3
torch==1.10.0
tqdm==4.64.0
wandb==0.12.9
